feat(v11.1.0): expose the §19.7.1.3 v3 producer via PyO3 (#328) — unblock the descent flag-day
Persist >= 16 fail-closes any pre-v3 tier at put_aggregated_tier (CIRISVerify#191
flag-day, reject token aggregation_meta_multiplicity) — and edge v11.0.0 shipped
the only correct v3 producer without exposing it to Python, so no descent driver
(CIRISConformance test_541 / server / agent) could produce an admissible tier.

## New Python surface (ungated — ships on every platform)
- `content_multiplicity(members, corpus_kind)` -> {member_masses,
  max_source_multiplicity}: the §19.7.1.3 measurement at fold time.
- `assemble_tier_meta_v3(...)` -> the full v3 meta dict + `signing_preimage`
  bytes (the exact §19.7.1 canonical bytes to sign before put_aggregated_tier).
  The FFI boundary pre-validates the Rust producer's panicking preconditions ->
  ValueError, never a panic across FFI.

## The producer moved out of the codec — load-bearing, not cosmetic
`content_multiplicity` needs NO codec: it is pure byte math + resampling. It only
LIVED in `realtime_av_codec::fountain` because that is where the fold is. Gating
the wheel on `codec-fountain` broke the **android armeabi-v7a** lane: raptorq
2.0.1 uses ARM NEON intrinsics (`uint8x16_t`, `stdarch_arm_neon_intrinsics`) that
are unstable on 32-bit ARM. So the surface now lives in `holonomic::multiplicity`
(ungated, a sibling of the `aggregation` producer where it semantically belongs),
and the wheel feature sets pull NO raptorq — verified: `cargo tree --features
pyo3 -i raptorq` finds nothing, while both pyfunctions stay registered.

`resample_nearest` is now a SINGLE shared definition used by both the fold and
the similarity measurement. That sharing is a correctness requirement, not a
convenience: similarity must be measured on exactly the content the fold
collapsed — two definitions could silently drift and fork a SIGNED field.

682 lib tests (incl. the R9 case now running with no codec feature at all);
clippy -D warnings clean across default / codec-fountain / transport-reticulum /
pyo3 / pyo3+reticulum+codec. Closes #328.
